refactor: Cleanup of the notification module
